+++ src/main/java/com/project/airbnb_app/controller/HotelBookingController.java
@@ -4,6 +4,7 @@ import com.project.airbnb_app.dto.GuestDto;
 import com.project.airbnb_app.dto.HotelBookingDto;
 import com.project.airbnb_app.dto.request.GuestCreateDto;
 import com.project.airbnb_app.dto.request.HotelBookingRequest;
+import com.project.airbnb_app.service.HotelBookingOrchestratorService;
 import com.project.airbnb_app.service.HotelBookingService;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
@@ -21,12 +22,13 @@ import java.util.Map;
 @Tag(name = "Hotel Booking API")
 public class HotelBookingController {
 
+    private final HotelBookingOrchestratorService hotelBookingOrchestratorService;
     private final HotelBookingService hotelBookingService;
 
     @PostMapping("/{bookingId}/guests")
     ResponseEntity<List<GuestDto>> addGuestsToBooking(@PathVariable Long bookingId, @RequestBody GuestCreateDto guestCreate) {
         return new ResponseEntity<>(
-                hotelBookingService.addGuestsToBooking(
+                hotelBookingOrchestratorService.addGuestsToBooking(
                         bookingId,
                         guestCreate.getGuest()
                 ),
@@ -36,13 +38,13 @@ public class HotelBookingController {
 
     @PatchMapping("/{bookingId}/cancel")
     ResponseEntity<Map<String, String>> cancelBooking(@PathVariable Long bookingId) {
-        hotelBookingService.cancelBooking(bookingId);
+        hotelBookingOrchestratorService.cancelBooking(bookingId);
         return ResponseEntity.ok(Map.of("status", String.format("Booking cancelled with th bookingId: %s", bookingId)));
     }
 
     @PostMapping
     public ResponseEntity<HotelBookingDto> createBooking(@Valid @RequestBody HotelBookingRequest hotelBookingRequest) {
-        HotelBookingDto hotelBookingDto = hotelBookingService.createHotelBooking(hotelBookingRequest);
+        HotelBookingDto hotelBookingDto = hotelBookingOrchestratorService.createHotelBooking(hotelBookingRequest);
         return new ResponseEntity<>(hotelBookingDto, HttpStatus.CREATED);
     }
 
@@ -53,7 +55,7 @@ public class HotelBookingController {
 
     @GetMapping("/{bookingId}/initPayment")
     public ResponseEntity<Map<String, String>> initiatePayment(@PathVariable Long bookingId) {
-        String paymentSessionUrl = hotelBookingService.initiatePayment(bookingId);
+        String paymentSessionUrl = hotelBookingOrchestratorService.initiatePayment(bookingId);
         return ResponseEntity.ok(Map.of("paymentSessionUrl", paymentSessionUrl));
     }
 }
