+++ src/main/java/com/project/airbnb_app/service/HotelBookingOrchestratorService.java
@@ -1,8 +1,18 @@
 package com.project.airbnb_app.service;
 
-import com.project.airbnb_app.entity.HotelBooking;
+import com.project.airbnb_app.dto.GuestDto;
+import com.project.airbnb_app.dto.HotelBookingDto;
+import com.project.airbnb_app.dto.request.HotelBookingRequest;
+
+import java.util.List;
 
 public interface HotelBookingOrchestratorService {
 
-    void save(HotelBooking hotelBooking);
+    List<GuestDto> addGuestsToBooking(Long bookingId, List<GuestDto> guestDtoList);
+
+    void cancelBooking(Long bookingId);
+
+    HotelBookingDto createHotelBooking(HotelBookingRequest hotelBookingRequest);
+
+    String initiatePayment(Long bookingId);
 }
